+++ src/stores/frame.store.ts
@@ -1,7 +1,7 @@
 import { writable } from "svelte/store";
 import { ArduinoFrameContainer } from "../core/frames/arduino.frame";
 import { MicroControllerType } from "../core/microcontroller/microcontroller";
-import { defaultSetting } from "./settings.store";
+import { defaultSetting } from "../firebase/model";
 
 const stateStore = writable<ArduinoFrameContainer>({
   frames: [],
