+++ src/stores/settings.store.ts
@@ -1,40 +1,34 @@
 import { writable } from 'svelte/store';
+import { Settings } from '../firebase/model';
 import is_browser from '../helpers/is_browser';
+import { defaultSetting } from '../firebase/model';
+import _ from 'lodash';
+import authStore from './auth.store';
 
-export interface Settings {
-  backgroundColor: string;
-  touchSkinColor: string;
-  ledColor: string;
-  customLedColor: boolean;
-  maxTimePerMove: number;
-}
-
-export const defaultSetting: Settings = {
-        backgroundColor: "#d9e4ec",
-        touchSkinColor: "#a424d3",
-        ledColor: "#AA0000",
-        customLedColor: false,
-        maxTimePerMove: 1000,
-      }
 
 const settings =
   is_browser() && localStorage.getItem("settings")
     ? JSON.parse(localStorage.getItem("settings"))
     : defaultSetting;
 
-// creates the store
 const settingsStore = writable<Settings>(settings);
 
-// updates local store with the new settings
-settingsStore.subscribe((settings) => {
-  if (is_browser()) {
-    localStorage.setItem('settings', JSON.stringify(settings));
+
+authStore.subscribe(async auth => {
+  if (auth.isLoggedIn && is_browser()) {
+    const db = firebase.firestore();
+    const settings = await db.collection('settings').doc(auth.uid).get();
+    settingsStore.set(settings.data());
   }
 });
 
-
+settingsStore.subscribe(newSettings => {
+  if (is_browser()) {
+    localStorage.setItem('settings', JSON.stringify(newSettings));
+  }
+})
 
 export default {
   subscribe: settingsStore.subscribe,
-  set: settingsStore.set,
+  set: settingsStore.set
 };
